+++ modules/features/taskerplugin/src/main/java/au/com/shiftyjelly/pocketcasts/taskerplugin/base/ActivityConfigBase.kt
@@ -3,7 +3,7 @@ package au.com.shiftyjelly.pocketcasts.taskerplugin.base
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.Composable
+import androidx.compose.material.Text
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.taskerplugin.base.hilt.appTheme
 
@@ -14,10 +14,24 @@ abstract class ActivityConfigBase<TViewModel : ViewModelBase<*, *>> : ComponentA
         viewModel.onCreate({ finish() }, { intent }, { code, data -> setResult(code, data) })
         setContent {
             AppThemeWithBackground(themeType = appTheme.activeTheme) {
-                Content()
+                val taskerVariables = viewModel.taskerVariables
+                val inputs = viewModel.inputFields.map { field ->
+                    TaskerInputFieldState.Content(
+                        field.valueState,
+                        field.labelResId,
+                        field.iconResId,
+                        field.shouldAskForState,
+                        { field.value = it },
+                        taskerVariables,
+                        field.getPossibleValues()
+                    ) {
+                        Text(field.getValueDescription(it))
+                    }
+                }
+                ComposableTaskerInputFieldList(inputs) {
+                    viewModel.finishForTasker()
+                }
             }
         }
     }
-    @Composable
-    protected abstract fun Content()
 }